포스트 페이징 버그 수정

+++ drawmind-server/src/main/java/com/example/codemindprojbackend/api/request/PostRequest.java
@@ -38,7 +38,7 @@ public class PostRequest {
     @ToString
     @RequiredArgsConstructor
     public static class List {
-        private Long currentPage;
+        private int currentPage;
     }
 
     @Getter
